class_work 05: create class-dir for classes, fix project imports, edit user-class
