+++ classes/Notifications.py
@@ -1,38 +1,17 @@
-from .sql_class import SqlConnection, Vacancy
 from aiogram.exceptions import TelegramRetryAfter, TelegramForbiddenError
 from aiogram import Bot
+
+from .SqlConnection import SqlConnection
+from .Users import User
+from .Vacancies import Vacancy
+from . import _sql_connection
+
 import asyncio
 
 
-class User:
-    def __init__(self, tg_id: int, username: str = None, fullname: str= None):
-        self.tg_id = tg_id
-        self.username = username
-        self.fullname = fullname
-
-
-class UserCommands:
-    def __init__(self, sql_connection: SqlConnection):
-        self.sql_conn = sql_connection
-
-    async def add_to_db(self, user: User):
-        # сделать возможнсть получать из аргумента пользователя которому тд и тп
-        self.sql_conn.cur.execute("INSERT OR IGNORE "
-                                  "INTO users (tg_id, username, fullname, active) "
-                                  "VALUES (?, ?, ?, ?)",
-                                  (user.tg_id, user.username, user.fullname, 1))
-
-        self.sql_conn.cur.execute("UPDATE users "
-                                  "SET username = ?, fullname = ?, active = ? "
-                                  "WHERE tg_id = ?",
-                                  (user.username, user.fullname, 1, user.tg_id))
-
-        self.sql_conn.conn.commit()
-
-
 class VacancyNotification:
-    def __init__(self, sql_connection: SqlConnection):
-        self.sql_conn = sql_connection
+    def __init__(self):
+        self.sql_conn: SqlConnection = _sql_connection
 
     async def check_existing_table(self, name_table: int) -> bool:
         self.sql_conn.cur.execute("SELECT name "
