+++ classes/Vacancies.py
@@ -1,62 +1,10 @@
-import sqlite3
+from .SqlConnection import SqlConnection
+from .RedisCommands import RedisCommands
+from .DataBaseCommands import DatabaseCommands
+from .Users import User
+
 from datetime import datetime
-import redis
 from assets import texts
-from user_class import User
-
-columns_titles = ["id", "employer", "work_type", "salary", "min_age", "min_exp", "datetime", "s_dscr", "l_dscr"]
-
-
-class SqlConnection:
-
-    def __init__(self, cur: sqlite3, conn: sqlite3):
-        self.cur: sqlite3 = cur
-        self.conn: sqlite3 = conn
-
-    async def close_conn(self):
-        self.conn.close()
-
-
-class DatabaseCommands:
-    def __init__(self, sql_connection: SqlConnection):
-        self.sql_conn: SqlConnection = sql_connection
-
-    async def get_row_by_id(self, row_id_in_db: int) -> tuple:
-        # Получение строки из бд по передаваемому id
-
-        self.sql_conn.cur.execute("SELECT * FROM vacancies WHERE id = ?", (row_id_in_db,))
-        row_in_db: tuple = self.sql_conn.cur.fetchone()
-        return row_in_db
-
-    @staticmethod
-    async def row_to_dict(row_in_db: tuple) -> dict:
-        # Переделывание передаваемой в функцию строки из бд в словарь
-
-        diction: dict = {}
-        for i, column in enumerate(columns_titles):
-            diction[column] = row_in_db[i]
-
-        return diction
-
-    @staticmethod
-    async def dict_to_text(vacancy_values: dict, type_descr: str) -> str:
-        employer: str = vacancy_values['employer']
-        work_type: str = vacancy_values['work_type']
-        salary: str = vacancy_values['salary']
-        min_age: str = f"Минимальный возраст: {vacancy_values['min_age']}\n" if vacancy_values['min_age'] is not None else ""
-        min_exp: str = f"Минимальный опыт работы: {vacancy_values['min_exp']}\n" if vacancy_values['min_exp'] is not None else ""
-        datetime: str = vacancy_values['datetime']
-        descr: str = vacancy_values['s_dscr'] if type_descr == "short" else vacancy_values['l_dscr']
-
-        final_text = (f"<b>{employer}</b>\n"
-                      f"{work_type}\n"
-                      f"{salary}\n"
-                      f"{min_age}"
-                      f"{min_exp}"
-                      f"{datetime}\n"
-                      f"{descr}")
-
-        return final_text
 
 
 class Vacancy:
@@ -68,31 +16,6 @@ class Vacancy:
         self.values: dict = values
         self.id: int = id
         self.text: str = text
-
-
-class RedisCommands:
-    def __init__(self):
-        self.redis_client: redis.Redis = redis.Redis(host='localhost', db=1, charset="utf-8", decode_responses=True)
-
-    async def close_conn(self):
-        self.redis_client.close()
-
-    async def user_add_history(self, user: User, vacancy: Vacancy):
-        try:
-            await self.redis_client.sadd(f"{user.tg_id}_history", vacancy.id)
-            await self.redis_client.expire(f"{user.tg_id}_history", 86400)
-            return True
-
-        except Exception as ex:
-            return False
-
-    async def user_get_history(self, user: User) -> set | bool:
-        history = self.redis_client.smembers(f"{user.tg_id}_history")
-
-        if history:
-            return history
-        else:
-            return False
 
 
 class VacanciesCommands:
@@ -127,7 +50,7 @@ class VacanciesCommands:
         except Exception as ex:
             return False
 
-    async def --------------------to_text(self, vacancy: Vacancy, type_descr: str) -> str:
+    async def to_text(self, vacancy: Vacancy, type_descr: str) -> str:
         # !!!!!!!!!!!!!! Почему бы не работать сразу с values вакансии, нежели с id
 
         if vacancy.id:  # если у вакансии id, а не ряд
