--- handlers/employ/__init__.py
@@ -1,1 +1,0 @@
-from . import view_vacancies
