--- handlers/employer/__init__.py
@@ -1,1 +1,0 @@
-from . import create_vacancy, edit_vacancy, error_processing
