+++ handlers/employer/edit_vacancy.py
@@ -1,23 +1,20 @@
-from aiogram.types import Message, CallbackQuery
-from aiogram.fsm.context import FSMContext
-from fsm.statesform import StapesForm as sf
-from assets import texts
+from aiogram.types import CallbackQuery
 from aiogram import Router, F, Bot
 from aiogram.filters import Text, StateFilter, Command
-
-from keyboards.inline_keyboards import inkb_edit_cancel_save
-
 from aiogram.types import Message
 from aiogram.fsm.context import FSMContext
+
 from fsm.statesform import StapesForm as sf
+
 from keyboards.inline_keyboards import *
 from keyboards.inline_keyboards import create_inkb
 
+from classes import db_commands
+from assets import texts
 import asyncio
-from objects import db_commands
+
 
 router = Router()
-
 router.callback_query.filter(StateFilter(sf.confirm_create))
