+++ handlers/main_window/main_page.py
@@ -1,12 +1,8 @@
-import asyncio
-
-from aiogram.types import Message, CallbackQuery
-from aiogram.fsm.context import FSMContext
+from aiogram.types import Message
 from aiogram.fsm.state import default_state
-from fsm.statesform import StapesForm as sf
 from assets import texts
-from aiogram import Router, Bot, F
-from aiogram.filters import Command, Text, StateFilter
+from aiogram import Router
+from aiogram.filters import Command, StateFilter
 
 from keyboards.inline_keyboards import *
