--- methods/__init__.py
@@ -1,1 +1,0 @@
-from .message import send_preview
